+++ src/drizzle-orm/index.js
@@ -1,23 +1,24 @@
 const { sql } = require('drizzle-orm');
 const { bit } = require('./bit');
 const { halfvec } = require('./halfvec');
+const { sparsevec } = require('./sparsevec');
 const { vector } = require('./vector');
 const utils = require('../utils');
 
 function l2Distance(column, value) {
-  return sql`${column} <-> ${utils.toSql(value)}`;
+  return sql`${column} <-> ${utils.toAnySql(value)}`;
 }
 
 function maxInnerProduct(column, value) {
-  return sql`${column} <#> ${utils.toSql(value)}`;
+  return sql`${column} <#> ${utils.toAnySql(value)}`;
 }
 
 function cosineDistance(column, value) {
-  return sql`${column} <=> ${utils.toSql(value)}`;
+  return sql`${column} <=> ${utils.toAnySql(value)}`;
 }
 
 function l1Distance(column, value) {
-  return sql`${column} <+> ${utils.toSql(value)}`;
+  return sql`${column} <+> ${utils.toAnySql(value)}`;
 }
 
 function hammingDistance(column, value) {
@@ -32,6 +33,7 @@ module.exports = {
   vector,
   halfvec,
   bit,
+  sparsevec,
   l2Distance,
   maxInnerProduct,
   cosineDistance,