+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added support for `halfvec` and `sparsevec` types to Knex.js
 - Added support for `halfvec` type to Sequelize
 - Added support for `halfvec` and `bit` types to MikroORM
-- Added support for `halfvec` and `bit` types to Drizzle ORM
+- Added support for `halfvec`, `bit`, and `sparsevec` types to Drizzle ORM
 - Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Knex.js
 - Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Objection.js
 - Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Kysely